(#81): TourResultSwiper 역할분리
