+++ src/pages/tour/geotrip/components/TourBottomSheet.tsx
@@ -1,0 +1,53 @@
+import { BottomSheet, LoadingSpinner, TourCard } from '@/components';
+import type { TourSummary } from '../types';
+import { TourOverView } from './';
+import { Suspense } from 'react';
+import { useStartTrip } from '../lib';
+
+interface TourBottomSheetProps extends TourSummary {
+  isOpen: boolean;
+  onClose: () => void;
+}
+
+export default function TourBottomSheet({
+  title,
+  dist,
+  firstimage,
+  contenttypeid,
+  isOpen,
+  contentid,
+  mapx,
+  mapy,
+  onClose,
+}: TourBottomSheetProps) {
+  const { handleStartTrip } = useStartTrip();
+  return (
+    <BottomSheet isOpen={isOpen} onClose={onClose} showOverlay={false}>
+      <div className="bg-white w-full">
+        <TourCard
+          title={title}
+          distance={dist}
+          imgUrl={firstimage || ''}
+          tourTypeId={contenttypeid}
+        />
+        <Suspense fallback={<LoadingSpinner />}>
+          <TourOverView contentId={contentid} />
+        </Suspense>
+        <div className="mt-4 w-full flex items-center justify-center">
+          <button
+            type="button"
+            className="bg-gradient-to-r from-primary-orange to-primary-red rounded-[15px] w-[320px] h-[50px] text-black font-bold text-[16px] shadow-[0_4px_16px_0_rgba(250,129,47,0.3)]"
+            onClick={() => {
+              handleStartTrip({
+                lng: mapx,
+                lat: mapy,
+              });
+            }}
+          >
+            여행 시작하기
+          </button>
+        </div>
+      </div>
+    </BottomSheet>
+  );
+}
