+++ src/pages/tour/geotrip/components/index.ts
@@ -2,3 +2,4 @@ export { default as TourOverView } from './TourOverView';
 export { default as TourResultSwiper } from './TourResultSwiper';
 export { default as TourSlide } from './TourSlide';
 export { default as TourSlideImages } from './TourSlideImages';
+export { default as TourBottomSheet } from './TourBottomSheet';
