+++ src/pages/tour/geotrip/lib/index.ts
@@ -1,2 +1,3 @@
 export { default as useTTS } from './useTTS';
 export { default as useStartTrip } from './useStartTrip';
+export { default as useCurrentTourInfo } from './useCurrentTourInfo';
